techDebt(BURefactorWorkerLambda): Move workerHeaders data into separate file
